# swagger/monitoring-service-open-api.yml
openapi: 3.0.0
info:
  title: Monitoring Service API
  description: This API allows users to manage readings and meters
  version: 1.0.0
servers:
  - url: http://localhost:8080/SpringFrameworkApp_war_exploded
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '201':
          description: The registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid user data
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: Logged in
        '400':
          description: Invalid credentials
...
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        email:
          type: string
          format: email
        password:
          type: string
          format: password
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic